refactor: decompose 4 more monolith files into submodules (issue #2486 sprint 8)
- mesh_loader.py (860 → 194 LOC) → _mesh_types, _mesh_parsers
- simscape_converter.py (856 → 381 LOC) → _body_conversion, _joint_conversion,
  _graph_utils
- green_surface.py (854 → 151 LOC) → _surface_data, _surface_geometry,
  _surface_analysis, _surface_io, _surface_presets, _checkpoint
- grip_contact_model.py (833 → 65 LOC) → _contact_types, _friction_laws,
  _grip_model, _grip_forces, _grip_exporter

Also updates _dynamics.py and _sim_core.py in putting_green with related changes.

All public symbols remain importable from original module paths.
