+++ src/unreal_integration/mesh_loader.py
@@ -29,350 +29,43 @@ Usage:
 
 from __future__ import annotations
 
-import json
 import logging
-import struct
-from dataclasses import dataclass, field
-from enum import Enum
 from pathlib import Path
 
-import numpy as np
+from ._mesh_parsers import (
+    load_collada,
+    load_fbx,
+    load_gltf,
+    load_obj,
+    load_ply,
+    load_stl,
+)
+from ._mesh_types import (
+    LoadedMesh,
+    MeshBone,
+    MeshFace,
+    MeshFormat,
+    MeshLoadError,
+    MeshMaterial,
+    MeshSkeleton,
+    MeshVertex,
+    UnsupportedFormatError,
+)
+
+__all__ = [
+    "LoadedMesh",
+    "MeshBone",
+    "MeshFace",
+    "MeshFormat",
+    "MeshLoadError",
+    "MeshLoader",
+    "MeshMaterial",
+    "MeshSkeleton",
+    "MeshVertex",
+    "UnsupportedFormatError",
+]
 
 logger = logging.getLogger(__name__)
-
-
-class MeshLoadError(Exception):
-    """Exception raised when mesh loading fails."""
-
-    def __init__(
-        self, message: str, path: str | None = None, cause: Exception | None = None
-    ) -> None:
-        if not (message is not None):
-            raise ValueError("message must be provided")
-        if not (message is not None):
-            raise ValueError("message must be provided")
-        super().__init__(message)
-        self.path = path
-        self.cause = cause
-
-
-class UnsupportedFormatError(MeshLoadError):
-    """Exception raised when mesh format is not supported."""
-
-    def __init__(self, extension: str, path: str | None = None) -> None:
-        if not (extension is not None):
-            raise ValueError("extension must be provided")
-        if not (extension is not None):
-            raise ValueError("extension must be provided")
-        super().__init__(f"Unsupported mesh format: {extension}", path)
-        self.extension = extension
-
-
-class MeshFormat(Enum):
-    """Supported mesh file formats."""
-
-    OBJ = "obj"
-    STL = "stl"
-    GLTF = "gltf"
-    GLB = "glb"
-    FBX = "fbx"
-    COLLADA = "dae"
-    PLY = "ply"
-
-    @property
-    def extension(self) -> str:
-        """Get file extension for format."""
-        return f".{self.value}"
-
-    @classmethod
-    def from_extension(cls, ext: str) -> MeshFormat:
-        """Get format from file extension.
-
-        Args:
-            ext: File extension (e.g., ".obj").
-
-        Returns:
-            Corresponding MeshFormat.
-
-        Raises:
-            UnsupportedFormatError: If extension is not supported.
-        """
-        ext_lower = ext.lower().lstrip(".")
-        for fmt in cls:
-            if fmt.value == ext_lower:
-                return fmt
-        raise UnsupportedFormatError(ext)
-
-
-@dataclass
-class MeshVertex:
-    """Vertex data for a mesh.
-
-    Attributes:
-        position: 3D position (x, y, z).
-        normal: Normal vector (optional).
-        uv: Texture coordinates (optional).
-        color: Vertex color RGBA (optional).
-        bone_indices: Skeleton bone indices for skinning (optional).
-        bone_weights: Skeleton bone weights for skinning (optional).
-    """
-
-    position: np.ndarray
-    normal: np.ndarray | None = None
-    uv: np.ndarray | None = None
-    color: np.ndarray | None = None
-    bone_indices: np.ndarray | None = None
-    bone_weights: np.ndarray | None = None
-
-
-@dataclass
-class MeshFace:
-    """Face (polygon) data for a mesh.
-
-    Attributes:
-        indices: Vertex indices forming the face.
-        material_index: Index into materials array (optional).
-    """
-
-    indices: np.ndarray
-    material_index: int = 0
-
-    @property
-    def is_triangle(self) -> bool:
-        """Check if face is a triangle."""
-        return len(self.indices) == 3
-
-
-@dataclass
-class MeshMaterial:
-    """Material data for mesh rendering.
-
-    Supports PBR (Physically Based Rendering) workflow.
-
-    Attributes:
-        name: Material name.
-        base_color: RGBA base color.
-        metallic: Metallic factor (0-1).
-        roughness: Roughness factor (0-1).
-        emissive: Emissive color RGB.
-        base_color_texture: Path to base color texture.
-        normal_texture: Path to normal map texture.
-        metallic_roughness_texture: Path to metallic-roughness texture.
-        occlusion_texture: Path to ambient occlusion texture.
-        emissive_texture: Path to emissive texture.
-    """
-
-    name: str = "default"
-    base_color: tuple[float, float, float, float] = (0.8, 0.8, 0.8, 1.0)
-    metallic: float = 0.0
-    roughness: float = 0.5
-    emissive: tuple[float, float, float] = (0.0, 0.0, 0.0)
-    base_color_texture: str | None = None
-    normal_texture: str | None = None
-    metallic_roughness_texture: str | None = None
-    occlusion_texture: str | None = None
-    emissive_texture: str | None = None
-
-    @classmethod
-    def default(cls) -> MeshMaterial:
-        """Create default material."""
-        return cls(name="default")
-
-
-@dataclass
-class MeshBone:
-    """Bone data for skeletal meshes.
-
-    Attributes:
-        name: Bone name.
-        index: Bone index in skeleton.
-        parent_index: Parent bone index (-1 for root).
-        local_transform: Local transformation matrix (4x4).
-        inverse_bind_matrix: Inverse bind pose matrix (optional).
-    """
-
-    name: str
-    index: int
-    parent_index: int
-    local_transform: np.ndarray
-    inverse_bind_matrix: np.ndarray | None = None
-
-    @property
-    def is_root(self) -> bool:
-        """Check if bone is root (no parent)."""
-        return self.parent_index < 0
-
-
-@dataclass
-class MeshSkeleton:
-    """Skeleton data for skeletal meshes.
-
-    Attributes:
-        bones: List of bones in skeleton.
-    """
-
-    bones: list[MeshBone] = field(default_factory=list)
-
-    @property
-    def bone_count(self) -> int:
-        """Get number of bones."""
-        return len(self.bones)
-
-    @property
-    def bone_names(self) -> list[str]:
-        """Get list of all bone names."""
-        return [bone.name for bone in self.bones]
-
-    @property
-    def root_bone(self) -> MeshBone | None:
-        """Get root bone."""
-        for bone in self.bones:
-            if bone.is_root:
-                return bone
-        return None
-
-    def get_bone(self, name: str) -> MeshBone | None:
-        """Get bone by name.
-
-        Args:
-            name: Bone name to find.
-
-        Returns:
-            MeshBone if found, None otherwise.
-        """
-        if not (name is not None):
-            raise ValueError("name must be provided")
-        if not (name is not None):
-            raise ValueError("name must be provided")
-        for bone in self.bones:
-            if bone.name == name:
-                return bone
-        return None
-
-    def get_bone_by_index(self, index: int) -> MeshBone | None:
-        """Get bone by index.
-
-        Args:
-            index: Bone index to find.
-
-        Returns:
-            MeshBone if found, None otherwise.
-        """
-        if not (index is not None):
-            raise ValueError("index must be provided")
-        if not (index is not None):
-            raise ValueError("index must be provided")
-        for bone in self.bones:
-            if bone.index == index:
-                return bone
-        return None
-
-    def get_children(self, parent_index: int) -> list[MeshBone]:
-        """Get child bones of a bone.
-
-        Args:
-            parent_index: Index of parent bone.
-
-        Returns:
-            List of child bones.
-        """
-        return [bone for bone in self.bones if bone.parent_index == parent_index]
-
-
-@dataclass
-class LoadedMesh:
-    """Complete loaded mesh data.
-
-    Attributes:
-        name: Mesh name.
-        vertices: List of vertices.
-        faces: List of faces.
-        materials: List of materials.
-        skeleton: Skeleton data (optional).
-        source_path: Original file path.
-        format: Source file format.
-    """
-
-    name: str
-    vertices: list[MeshVertex]
-    faces: list[MeshFace]
-    materials: list[MeshMaterial] = field(default_factory=list)
-    skeleton: MeshSkeleton | None = None
-    source_path: str | None = None
-    format: MeshFormat | None = None
-
-    @property
-    def vertex_count(self) -> int:
-        """Get number of vertices."""
-        return len(self.vertices)
-
-    @property
-    def face_count(self) -> int:
-        """Get number of faces."""
-        return len(self.faces)
-
-    @property
-    def has_skeleton(self) -> bool:
-        """Check if mesh has skeleton data."""
-        return self.skeleton is not None and self.skeleton.bone_count > 0
-
-    @property
-    def has_normals(self) -> bool:
-        """Check if mesh has normal data."""
-        return any(v.normal is not None for v in self.vertices)
-
-    @property
-    def has_uvs(self) -> bool:
-        """Check if mesh has UV coordinates."""
-        return any(v.uv is not None for v in self.vertices)
-
-    @property
-    def bounding_box(self) -> dict[str, np.ndarray]:
-        """Calculate mesh bounding box.
-
-        Returns:
-            Dictionary with 'min' and 'max' bounds.
-        """
-        if not self.vertices:
-            return {"min": np.zeros(3), "max": np.zeros(3)}
-
-        positions = np.array([v.position for v in self.vertices])
-        return {
-            "min": np.min(positions, axis=0),
-            "max": np.max(positions, axis=0),
-        }
-
-    def to_arrays(self) -> tuple[np.ndarray, np.ndarray]:
-        """Convert to numpy arrays.
-
-        Returns:
-            Tuple of (positions, indices) arrays.
-        """
-        positions = np.array([v.position for v in self.vertices])
-        indices = np.array([f.indices for f in self.faces])
-        return positions, indices
-
-    def get_normals_array(self) -> np.ndarray | None:
-        """Get normals as numpy array.
-
-        Returns:
-            Normals array, or None if no normals.
-        """
-        if not self.has_normals:
-            return None
-        return np.array(
-            [v.normal if v.normal is not None else [0, 0, 0] for v in self.vertices]
-        )
-
-    def get_uvs_array(self) -> np.ndarray | None:
-        """Get UVs as numpy array.
-
-        Returns:
-            UVs array, or None if no UVs.
-        """
-        if not self.has_uvs:
-            return None
-        return np.array([v.uv if v.uv is not None else [0, 0] for v in self.vertices])
 
 
 class MeshLoader:
@@ -455,14 +148,11 @@ class MeshLoader:
         """
         path_obj = Path(path)
 
-        # Precondition: file exists
         if not path_obj.exists():
             raise FileNotFoundError(f"Mesh file not found: {path}")
 
-        # Precondition: format supported
         fmt = MeshFormat.from_extension(path_obj.suffix)
 
-        # Check cache
         if self.enable_cache:
             mtime = path_obj.stat().st_mtime
             if path in self._cache:
@@ -471,27 +161,25 @@ class MeshLoader:
                     logger.debug(f"Using cached mesh: {path}")
                     return cached_mesh
 
-        # Load based on format
         try:
             if fmt == MeshFormat.OBJ:
-                mesh = self._load_obj(path_obj)
+                mesh = load_obj(path_obj)
             elif fmt == MeshFormat.STL:
-                mesh = self._load_stl(path_obj)
+                mesh = load_stl(path_obj)
             elif fmt in (MeshFormat.GLTF, MeshFormat.GLB):
-                mesh = self._load_gltf(path_obj)
+                mesh = load_gltf(path_obj)
             elif fmt == MeshFormat.FBX:
-                mesh = self._load_fbx(path_obj)
+                mesh = load_fbx(path_obj)
             elif fmt == MeshFormat.COLLADA:
-                mesh = self._load_collada(path_obj)
+                mesh = load_collada(path_obj)
             elif fmt == MeshFormat.PLY:
-                mesh = self._load_ply(path_obj)
+                mesh = load_ply(path_obj)
             else:
                 raise UnsupportedFormatError(path_obj.suffix, path)
 
             mesh.source_path = path
             mesh.format = fmt
 
-            # Update cache
             if self.enable_cache:
                 self._cache[path] = (path_obj.stat().st_mtime, mesh)
 
@@ -504,357 +192,3 @@ class MeshLoader:
             raise
         except (RuntimeError, TypeError, ValueError) as e:
             raise MeshLoadError(f"Failed to load mesh: {e}", path, e) from e
-
-    def _load_obj(self, path: Path) -> LoadedMesh:
-        """Load OBJ format mesh.
-
-        Args:
-            path: Path to OBJ file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        if not (path is not None):
-            raise ValueError("path must be provided")
-        if not (path is not None):
-            raise ValueError("path must be provided")
-        vertices: list[MeshVertex] = []
-        faces: list[MeshFace] = []
-        positions: list[np.ndarray] = []
-        normals: list[np.ndarray] = []
-        uvs: list[np.ndarray] = []
-
-        with open(path) as f:
-            for line in f:
-                line = line.strip()
-                if not line or line.startswith("#"):
-                    continue
-
-                parts = line.split()
-                if not parts:
-                    continue
-
-                if parts[0] == "v":  # Vertex position
-                    positions.append(
-                        np.array([float(parts[1]), float(parts[2]), float(parts[3])])
-                    )
-                elif parts[0] == "vn":  # Vertex normal
-                    normals.append(
-                        np.array([float(parts[1]), float(parts[2]), float(parts[3])])
-                    )
-                elif parts[0] == "vt":  # Texture coordinate
-                    uvs.append(np.array([float(parts[1]), float(parts[2])]))
-                elif parts[0] == "f":  # Face
-                    face_indices = []
-                    for i in range(1, len(parts)):
-                        # Parse v/vt/vn format
-                        vertex_data = parts[i].split("/")
-                        v_idx = int(vertex_data[0]) - 1  # OBJ indices are 1-based
-
-                        vt_idx = None
-                        vn_idx = None
-                        if len(vertex_data) > 1 and vertex_data[1]:
-                            vt_idx = int(vertex_data[1]) - 1
-                        if len(vertex_data) > 2 and vertex_data[2]:
-                            vn_idx = int(vertex_data[2]) - 1
-
-                        # Create vertex
-                        vertex = MeshVertex(
-                            position=positions[v_idx],
-                            normal=(
-                                normals[vn_idx]
-                                if vn_idx is not None and vn_idx < len(normals)
-                                else None
-                            ),
-                            uv=(
-                                uvs[vt_idx]
-                                if vt_idx is not None and vt_idx < len(uvs)
-                                else None
-                            ),
-                        )
-                        vertices.append(vertex)
-                        face_indices.append(len(vertices) - 1)
-
-                    faces.append(MeshFace(indices=np.array(face_indices)))
-
-        return LoadedMesh(
-            name=path.stem,
-            vertices=vertices,
-            faces=faces,
-        )
-
-    def _load_stl(self, path: Path) -> LoadedMesh:
-        """Load STL format mesh.
-
-        Args:
-            path: Path to STL file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        if not (path is not None):
-            raise ValueError("path must be provided")
-        if not (path is not None):
-            raise ValueError("path must be provided")
-        vertices: list[MeshVertex] = []
-        faces: list[MeshFace] = []
-
-        # Check if binary or ASCII
-        with open(path, "rb") as f:
-            header = f.read(80)
-            is_binary = not header.strip().startswith(b"solid")
-
-        if is_binary:
-            with open(path, "rb") as f:
-                f.read(80)  # Skip header
-                num_triangles = struct.unpack("<I", f.read(4))[0]
-
-                for _ in range(num_triangles):
-                    # Normal (3 floats)
-                    normal = struct.unpack("<3f", f.read(12))
-                    normal_arr = np.array(normal)
-
-                    # Three vertices (3 floats each)
-                    face_indices = []
-                    for _ in range(3):
-                        vertex = struct.unpack("<3f", f.read(12))
-                        vertices.append(
-                            MeshVertex(
-                                position=np.array(vertex),
-                                normal=normal_arr,
-                            )
-                        )
-                        face_indices.append(len(vertices) - 1)
-
-                    faces.append(MeshFace(indices=np.array(face_indices)))
-
-                    # Attribute byte count (unused)
-                    f.read(2)
-        else:
-            # ASCII STL parsing
-            with open(path) as f:
-                current_normal = None
-                face_vertices: list[int] = []
-
-                for line in f:
-                    line = line.strip()
-                    if line.startswith("facet normal"):
-                        parts = line.split()
-                        current_normal = np.array(
-                            [float(parts[2]), float(parts[3]), float(parts[4])]
-                        )
-                    elif line.startswith("vertex"):
-                        parts = line.split()
-                        vertices.append(
-                            MeshVertex(
-                                position=np.array(
-                                    [float(parts[1]), float(parts[2]), float(parts[3])]
-                                ),
-                                normal=current_normal,
-                            )
-                        )
-                        face_vertices.append(len(vertices) - 1)
-                    elif line.startswith("endfacet"):
-                        if len(face_vertices) == 3:
-                            faces.append(MeshFace(indices=np.array(face_vertices)))
-                        face_vertices = []
-
-        return LoadedMesh(
-            name=path.stem,
-            vertices=vertices,
-            faces=faces,
-        )
-
-    def _load_gltf(self, path: Path) -> LoadedMesh:
-        """Load GLTF/GLB format mesh.
-
-        Args:
-            path: Path to GLTF/GLB file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        # Try to use trimesh if available
-        try:
-            import trimesh
-
-            scene = trimesh.load(str(path))
-
-            # Handle scene vs mesh
-            if isinstance(scene, trimesh.Scene):
-                # Combine all meshes
-                meshes = list(scene.geometry.values())
-                if not meshes:
-                    raise MeshLoadError("No meshes found in scene", str(path))
-                mesh_data = meshes[0]
-                for m in meshes[1:]:
-                    mesh_data = trimesh.util.concatenate([mesh_data, m])
-            else:
-                mesh_data = scene
-
-            # Convert to our format
-            vertices = [
-                MeshVertex(
-                    position=mesh_data.vertices[i],
-                    normal=(
-                        mesh_data.vertex_normals[i]
-                        if hasattr(mesh_data, "vertex_normals")
-                        else None
-                    ),
-                )
-                for i in range(len(mesh_data.vertices))
-            ]
-
-            faces = [MeshFace(indices=face) for face in mesh_data.faces]
-
-            return LoadedMesh(
-                name=path.stem,
-                vertices=vertices,
-                faces=faces,
-            )
-
-        except ImportError:
-            # Fallback: basic GLTF JSON parsing
-            return self._load_gltf_basic(path)
-
-    def _load_gltf_basic(self, path: Path) -> LoadedMesh:
-        """Basic GLTF loading without trimesh.
-
-        Args:
-            path: Path to GLTF file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        if path.suffix.lower() == ".glb":
-            raise MeshLoadError("GLB loading requires trimesh library", str(path))
-
-        with open(path) as f:
-            gltf = json.load(f)
-
-        # This is a simplified implementation
-        # Full GLTF loading requires proper accessor/buffer handling
-        logger.warning("Basic GLTF loading - some features may not be supported")
-
-        if "meshes" not in gltf or not gltf["meshes"]:
-            raise MeshLoadError("No meshes found in GLTF", str(path))
-
-        # Full GLTF loading requires accessor/buffer parsing which needs trimesh.
-        # Returning placeholder geometry would silently hide the failure from callers.
-        raise MeshLoadError(
-            "GLTF accessor/buffer parsing requires the trimesh library",
-            str(path),
-        )
-
-    def _load_fbx(self, path: Path) -> LoadedMesh:
-        """Load FBX format mesh.
-
-        Args:
-            path: Path to FBX file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        # Try to use trimesh/pyfbx if available
-        try:
-            import trimesh
-
-            mesh = trimesh.load(str(path))
-
-            if isinstance(mesh, trimesh.Scene):
-                meshes = list(mesh.geometry.values())
-                if not meshes:
-                    raise MeshLoadError("No meshes found in FBX", str(path))
-                mesh = meshes[0]
-
-            if not isinstance(mesh, trimesh.Trimesh):
-                raise ValueError(f"Expected trimesh.Trimesh, got {type(mesh).__name__}")
-            vertices = [
-                MeshVertex(position=mesh.vertices[i]) for i in range(len(mesh.vertices))
-            ]
-
-            faces = [MeshFace(indices=face) for face in mesh.faces]
-
-            return LoadedMesh(
-                name=path.stem,
-                vertices=vertices,
-                faces=faces,
-            )
-
-        except ImportError as e:
-            raise MeshLoadError(
-                "FBX loading requires trimesh library: pip install trimesh[easy]",
-                str(path),
-            ) from e
-
-    def _load_collada(self, path: Path) -> LoadedMesh:
-        """Load COLLADA (.dae) format mesh.
-
-        Args:
-            path: Path to COLLADA file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        try:
-            import trimesh
-
-            mesh = trimesh.load(str(path))
-
-            if isinstance(mesh, trimesh.Scene):
-                meshes = list(mesh.geometry.values())
-                if not meshes:
-                    raise MeshLoadError("No meshes found in COLLADA", str(path))
-                mesh = meshes[0]
-
-            if not isinstance(mesh, trimesh.Trimesh):
-                raise ValueError(f"Expected trimesh.Trimesh, got {type(mesh).__name__}")
-            vertices = [
-                MeshVertex(position=mesh.vertices[i]) for i in range(len(mesh.vertices))
-            ]
-
-            faces = [MeshFace(indices=face) for face in mesh.faces]
-
-            return LoadedMesh(
-                name=path.stem,
-                vertices=vertices,
-                faces=faces,
-            )
-
-        except ImportError as e:
-            raise MeshLoadError(
-                "COLLADA loading requires trimesh library", str(path)
-            ) from e
-
-    def _load_ply(self, path: Path) -> LoadedMesh:
-        """Load PLY format mesh.
-
-        Args:
-            path: Path to PLY file.
-
-        Returns:
-            Loaded mesh data.
-        """
-        try:
-            import trimesh
-
-            mesh = trimesh.load(str(path))
-            if not isinstance(mesh, trimesh.Trimesh):
-                raise ValueError(f"Expected trimesh.Trimesh, got {type(mesh).__name__}")
-
-            vertices = [
-                MeshVertex(position=mesh.vertices[i]) for i in range(len(mesh.vertices))
-            ]
-
-            faces = [MeshFace(indices=face) for face in mesh.faces]
-
-            return LoadedMesh(
-                name=path.stem,
-                vertices=vertices,
-                faces=faces,
-            )
-
-        except ImportError as e:
-            raise MeshLoadError(
-                "PLY loading requires trimesh library", str(path)
-            ) from e
